Russian Localization
I did it.
I spent about a week translating this mod into Russian. I used several online translators for this, because each gave out its own version, and from them I chose the most suitable one. Moreover, I did not just copy and paste, I tried to translate, trying to stick to the idea, without breaking puns (which is very difficult in some places, and, alas, somewhere this meaning disappeared, which is a shame).
The translation has been checked in the game and is fully functional.
I don't know why no one has translated it yet. On the Internet, I did not see a single person who completed the work to the end, everyone was talking only about what they were going to, and then there was silence. I am proud of my work, I admit that it may not be perfect and there may be inaccuracies and incorrect translation in some places, I confess, but if someone shows me the places where I am wrong and shows me how to do it correctly, I will only be glad.